update caption field
